OAuth2: Add Resource Ownership Strict Mode Check
Enforce Strict Mode when updating / saving OAuth2 credentials. At the
moment we force the check at the plugin level which is not ideal and
will be removed in the near future.

In addition the change include setting the correct error tag so we can
show the error message in context.