deprecation: E_STRICT Constant
This addresses the deprecation of the `E_STRICT` constant. The upstream
error level has been removed and now the constant is being deprecated.
Since we don't support PHP 5.x any longer and this is now
deprecated/removed this fully removes the last mention of the `E_STRICT`
constant. So long and farewell my friend.

Reference:
https://www.php.net/manual/en/errorfunc.constants.php#constant.e-strict