deprecation: fgetcsv/fputcsv Default Escape Param
This addresses the deprecation of using the default value for the
`escape` parameter for both `fputcsv()` and `fgetcsv()`. Essentially,
you can no longer use the default value and instead have to explicitly
pass it in. This adds the needed defaults for `escape` as well as
`separator` and `enclosure` where needed to avoid warnings and errors.

References:
https://www.php.net/manual/en/migration84.deprecated.php#migration84.deprecated.standard
https://www.php.net/manual/en/function.fgetcsv.php#refsect1-function.fgetcsv-parameters